Backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.74
numpy==1.24.3
pymongo==4.4.1
google-generativeai==0.1.0
werkzeug==2.3.6
Pillow==10.0.0
python-dotenv==1.0.0